Move `types` module to `concepts`
This seems suitable when looking at the dependency diagram;
Just as actual type definitions of new datastructures, the

- type aliases
- NewType definitions
- TypedDict definitions

present in the `types` module are all of „leaf“-type in the dependency
digraph, and don't have any state.
Thus, I think they are suitable for this module.
